+++ SourceCode/GPS/Forms/FormSettings.cs
@@ -511,7 +511,7 @@ namespace AgOpenGPS
             mf.vehicle.slowSpeedCutoff = cutoffSpeed * cutoffMetricImperial;
 
             Properties.Settings.Default.setDisplay_lineWidth = (int)(nudLineWidth.Value);
-            mf.ABLine.lineWidth = (float)(nudLineWidth.Value);
+            mf.ABLine.lineWidth = (int)(nudLineWidth.Value);
 
             mf.tram.abOffset = (Math.Round((mf.vehicle.toolWidth - mf.vehicle.toolOverlap) / 2.0, 3));
